Load profile view based on provider
